Native: make range for Instants and LocalDate a 1000 times narrower
This is done to support `LocalDate.daysUntil(...): Int` always
having a precise value. Additionally, it allows to simplify some
parts of the implementation.

Before, the Native implementation was using the same representable
ranges as the Java implementation; now, it uses the same ranges as
the JS implementation, which are narrower so that everything fits
in a JS number.

+++ core/nativeMain/src/TimeZone.kt
@@ -107,7 +107,7 @@ public actual open class TimeZone internal constructor(private val tzid: TZID, a
             throw RuntimeException("Unable to acquire the offset at ${this@atZone} for zone ${this@TimeZone}")
         }
         val dateTime = try {
-            this@atZone.plusSeconds(transitionDuration.toLong())
+            this@atZone.plusSeconds(transitionDuration)
         } catch (e: IllegalArgumentException) {
             throw DateTimeArithmeticException("Overflow whet correcting the date-time to not be in the transition gap", e)
         } catch (e: ArithmeticException) {
